scripts/requirements.txt
duckduckgo-search>=7.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
